Extract Phoenix.Socket.Transport.dispatch to separate module
